Move all deprecated SimpleRedux APIs to a single file
