+++ packages/@misk/simpleredux/src/index.ts
@@ -1,22 +1,7 @@
 export * from "./action"
-import {
-  dispatchSimpleRedux,
-  IDispatchSimpleRedux,
-  // Deprecated
-  dispatchSimpleForm,
-  IDispatchSimpleForm,
-  dispatchSimpleNetwork,
-  IDispatchSimpleNetwork
-} from "./dispatch"
-export {
-  dispatchSimpleRedux,
-  IDispatchSimpleRedux,
-  // Deprecated
-  dispatchSimpleForm,
-  IDispatchSimpleForm,
-  dispatchSimpleNetwork,
-  IDispatchSimpleNetwork
-}
+export * from "./deprecated"
+import { dispatchSimpleRedux, IDispatchSimpleRedux } from "./dispatch"
+export { dispatchSimpleRedux, IDispatchSimpleRedux }
 export * from "./saga"
 export * from "./reducer"
 export * from "./utilities"
